修正：功能 - 异步发送消息通知消息时，VirtualWnd失效的问题

+++ DuiLib/Control/UIMenu.cpp
@@ -636,7 +636,7 @@ namespace DuiLib {
 	CMenuElementUI::CMenuElementUI():
 	m_pWindow(NULL),
 		m_bDrawLine(false),
-		m_dwLineColor(DEFAULT_LINE_COLOR),
+		m_dwLineColor((DWORD)DEFAULT_LINE_COLOR),
 		m_bCheckItem(false),
 		m_bShowExplandIcon(false)
 	{
